// resources/js/api/generated/models/containerStats.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Berth API
 * Berth: Opinionated docker compose stack management API
 * OpenAPI spec version: 1.0.0
 */

export interface ContainerStats {
  /** @minimum 0 */
  block_read_bytes: number;
  /** @minimum 0 */
  block_read_ops: number;
  /** @minimum 0 */
  block_write_bytes: number;
  /** @minimum 0 */
  block_write_ops: number;
  cpu_percent: number;
  /** @minimum 0 */
  cpu_system_time: number;
  /** @minimum 0 */
  cpu_user_time: number;
  /** @minimum 0 */
  memory_cache: number;
  /** @minimum 0 */
  memory_limit: number;
  memory_percent: number;
  /** @minimum 0 */
  memory_rss: number;
  /** @minimum 0 */
  memory_swap: number;
  /** @minimum 0 */
  memory_usage: number;
  name: string;
  /** @minimum 0 */
  network_rx_bytes: number;
  /** @minimum 0 */
  network_rx_packets: number;
  /** @minimum 0 */
  network_tx_bytes: number;
  /** @minimum 0 */
  network_tx_packets: number;
  /** @minimum 0 */
  page_faults: number;
  /** @minimum 0 */
  page_major_faults: number;
  service_name: string;
}