// resources/js/api/generated/models/stackStats.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Berth API
 * Berth: Opinionated docker compose stack management API
 * OpenAPI spec version: 1.0.0
 */
import type { ContainerStats } from './containerStats';
import type { HostStats } from './hostStats';

export interface StackStats {
  collected_at: string;
  containers: ContainerStats[];
  host: HostStats;
  /** @nullable */
  sample_window_seconds: number | null;
  stack_name: string;
}